+++ kamehouse-ui/src/main/webapp/js/admin/server-management.js
@@ -7,7 +7,6 @@
 class ServerManager {
 
   static #ADMIN_API_URL = "/kame-house-admin/api/v1/admin";
-  static #ALT_TAB_URL = '/screen/alt-tab-key-press';
   static #SUSPEND_URL = '/power-management/suspend';
   static #SHUTDOWN_URL = '/power-management/shutdown';
   static #REBOOT_URL = '/power-management/reboot';
@@ -30,20 +29,6 @@ class ServerManager {
     });
   }
   
-  /** Send ALT+TAB key press */
-  altTabCommand() {
-    const numberOfTabs = document.getElementById("number-of-tabs-dropdown").value;
-    kameHouse.logger.trace("Number of tab presses: " + numberOfTabs);
-    const requestParam = {
-      "tabs" : numberOfTabs
-    };
-    kameHouse.plugin.modal.loadingWheelModal.open();
-    const config = kameHouse.http.getConfig();
-    kameHouse.plugin.debugger.http.post(config, ServerManager.#ADMIN_API_URL + ServerManager.#ALT_TAB_URL, kameHouse.http.getUrlEncodedHeaders(), requestParam, 
-    (responseBody, responseCode, responseDescription, responseHeaders) => {this.#processSuccessSuspend(responseBody, responseCode, responseDescription, responseHeaders)}, 
-    (responseBody, responseCode, responseDescription, responseHeaders) => {this.#processErrorSuspend(responseBody, responseCode, responseDescription, responseHeaders)});
-  }
-
   /**
    * WakeOnLan
    */
@@ -176,19 +161,6 @@ class ServerManager {
     kameHouse.plugin.debugger.http.post(config, ServerManager.#ADMIN_API_URL + ServerManager.#HTTPD_URL, null, null, 
       (responseBody, responseCode, responseDescription, responseHeaders) => {this.#processSuccessHttpdRestart(responseBody, responseCode, responseDescription, responseHeaders)}, 
       (responseBody, responseCode, responseDescription, responseHeaders) => {this.#processErrorHttpdRestart(responseBody, responseCode, responseDescription, responseHeaders)});
-  }
-
-  /**
-   * Execute post http request.
-   */
-  post(url, requestBody) {
-    kameHouse.plugin.modal.loadingWheelModal.open();
-    let headers = null;
-    if (!kameHouse.core.isEmpty(requestBody)) {
-      headers = kameHouse.http.getApplicationJsonHeaders();
-    }
-    const config = kameHouse.http.getConfig();
-    kameHouse.plugin.debugger.http.post(config, ServerManager.#ADMIN_API_URL + url, headers, requestBody, this.#processSuccess, this.#processError);
   }
 
   /**