+++ changelog.md
@@ -12,7 +12,8 @@
 - Refactored all php code to use classes
 - Updated vlc player connection error log level
 - Added button for create all audio playlists in groot
-- Added ESC, ENTER and ALT+TAB key press buttons in admin server management page
+- Moved screen controller buttons to a separate page from server management
+- Added ESC, ENTER and ALT+TAB key press buttons in the screen controller page
 - Updated jvncsender to version 1.02
 - Fixed broken execute on docker host commands
 - Fixed bugs